Use 'super::#ident' path in `enumflags_derive`
Previously, `enumflags_derive` referenced the original type
through a plain '#ident' path. Since all of the generated code
is inside a new module, the result was something like this:

```rust
struct MyTypeWithDerive;

mod __scope_enumderive__MyTypeWithDerive {
    impl std::ops::Not for MyTypeWithDerive {
        ....
    }
}
```

Normally, code like this would not compile - the type `MyTypeWithDerive`
isn't imported in the new module. However, Rust contains a
backwards-compatibility hack to this to compile *only* when the
code is emitted by a derive macro.

The rustc project is trying to remove this hack from the compiler -
it increases the complexity of both the language and the compiler
implementation, and only exists to keep code from breaking.

This commit ensures that `enumflags` will continue to compile
on future versions of rustc (as well as existing version of rustc).
We now emit a `use super::#ident;` statement, resulting in code
that looks like this:

```rust
struct MyTypeWithDerive;

mod __scope_enumderive__MyTypeWithDerive {
    use super::MyTypeWithDerive;
    impl std::ops::Not for MyTypeWithDerive {
        ....
    }
}
```

This code is accepted by current versions of rustc, and will
continue to compile if the backwards-compatibility hack is removed
in a future version of rustc.